feat: move password changing functionality into `account`
